+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 ## [0.6.0] - 2023-05-24
 
+### Added
+
+- Added Codeberg as a recognised forge for the "forge quick view".
+
+### Changed
+
 - Relaxed the required [Textual](https://github.com/Textualize/textual)
   dependency version requirement.